Address Chris' review comment on a separate class for the config handler
The former `ZclWarningDeviceConfig` class is now inlined into the
`ZigBeeConverterWarningDevice` class.
